new exceptions, MissingSegmentError, MissingTagError
MissingSegmentError is raised when a segment can't be located.
In this commit, it is raised when dynamic segment is not found.
MissingTagError is raised when dynamic can't be located.

PatcherError is a catchall exception,
It creates problem for who use patchelf.rb as library.
Having a little bit of distinction helps, for e.g, when
quering runpath, if the ELF is corrupt the user may wish
perform some actions, but if its just runpath missing they
can choose to ignore.