Fix file descriptor leak
There are several places where `File.open` is called without ensuring
that the file is closed afterwards. In programs that handle many files,
this can lead to troubles when `ulimit -n` is set to a low value. For
example, in Homebrew, we have seen at least the following issues:

- https://github.com/Homebrew/brew/issues/19177
- https://github.com/Homebrew/brew/issues/19866
- https://github.com/Homebrew/brew/issues/20223
- https://github.com/Homebrew/brew/issues/20302

Fix the file descriptor leak by ensuring that the file handles are
properly closed after use.

Signed-off-by: Ruoyu Zhong <zhongruoyu@outlook.com>